renamed infodemand route, cleaned function up
